pyproject.toml
[tool.poetry]
name = "fastapi-enterprise-clean"
version = "0.1.0"
description = "Enterprise-grade FastAPI scaffold (Clean Architecture, SOLID, monitoring, Celery, Redis, gRPC)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic = "^2.7"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0"
alembic = "^1.13"
redis = "^5.0"
celery = "^5.3"
python-json-logger = "^2.0.7"
prometheus-client = "^0.20.0"
ldap3 = "^2.9"
httpx = "^0.27"
grpcio = "^1.64.0"
protobuf = "^5.27.0"
structlog = "^24.1.0"
PyYAML = "^6.0.0"
aiosqlite = "^0.21.0"
greenlet = "^3.2.3"
python-jose = "^3.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
mypy = "^1.10.0"
ruff = "^0.5.5"
black = "^24.4.2"
pre-commit = "^3.7.0"
bandit = "^1.7.7"
types-PyYAML = "^6.0.12"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "N", "S", "C90", "A", "PT", "RET", "SIM", "PL", "TRY", "FLY", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
warn_return_any = true
no_implicit_optional = true
strict_optional = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
